+++ components/raftstore-v2/tests/failpoints/mod.rs
@@ -5,4 +5,8 @@
 #![feature(custom_test_frameworks)]
 #![test_runner(test_util::run_failpoint_tests)]
 
+#[allow(dead_code)]
+#[path = "../integrations/cluster.rs"]
+mod cluster;
+mod test_basic_write;
 mod test_bootstrap;
