+++ components/raftstore-v2/tests/integrations/test_basic_write.rs
@@ -12,7 +12,7 @@ use raftstore::store::{INIT_EPOCH_CONF_VER, INIT_EPOCH_VER};
 use raftstore_v2::router::PeerMsg;
 use tikv_util::store::new_peer;
 
-use crate::Cluster;
+use crate::cluster::Cluster;
 
 /// Test basic write flow.
 #[test]
